requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
decorator==5.1.0
Django==3.2.6
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.7
fonttools==4.28.2
hvac==0.11.2
idna==3.2
kiwisolver==1.3.2
knox==0.1.14
loguru==0.5.3
matplotlib==3.5.0
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pycparser==2.20
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
validators==0.18.2
win32-setctime==1.0.3
